Moved sverdle Template in gen-craft
