+++ src/routes/gen-craft/+page.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { confetti } from '@neoconfetti/svelte';
 	import { enhance } from '$app/forms';
-	import type { PageData, ActionData } from './$types';
+	import type { PageData, ActionData } from '../sverdle/$types';
 	import { reduced_motion } from './reduced-motion';
 
 	export let data: PageData;
