+++ src/routes/gen-craft/game.ts
