+++ src/routes/gen-craft/how-to-play/+page.svelte
